+++ CHANGELOG.md
@@ -5,6 +5,7 @@ See [UniteJS Engine Changelog](https://github.com/unitejs/engine/blob/master/CHA
 ## 1.6.1
 
 * Added pwa (Progressive Web App) buildConfiguration option
+* Added version check after running commands
 
 ## 1.6.0